Consolidate makeBucketSpace() test functions to document module.
