+++ persistence/src/vespa/persistence/spi/test.cpp
@@ -1,33 +1,17 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include "test.h"
+#include <vespa/document/test/make_bucket_space.h>
 
 using document::BucketId;
 using document::BucketSpace;
+using document::test::makeBucketSpace;
 
 namespace storage::spi::test {
 
-BucketSpace makeBucketSpace()
-{
-    return BucketSpace::placeHolder();
-}
-
-BucketSpace makeBucketSpace(const vespalib::string &docTypeName)
-{
-    // Used by persistence conformance test to map fron document type name
-    // to bucket space.  See document::TestDocRepo for known document types.
-    if (docTypeName == "no") {
-        return BucketSpace(2);
-    } else if (docTypeName == "testdoctype2") {
-        return BucketSpace(1);
-    } else {
-        return makeBucketSpace();
-    }
-}
-
 Bucket makeSpiBucket(BucketId bucketId, PartitionId partitionId)
 {
-    return Bucket(document::Bucket(BucketSpace::placeHolder(), bucketId), partitionId);
+    return Bucket(document::Bucket(makeBucketSpace(), bucketId), partitionId);
 }
 
 Bucket makeSpiBucket(BucketId bucketId)
