+++ persistence/src/vespa/persistence/spi/test.h
@@ -8,8 +8,6 @@ namespace storage::spi::test {
 
 // Helper functions used by unit tests
 
-document::BucketSpace makeBucketSpace();
-document::BucketSpace makeBucketSpace(const vespalib::string &docTypeName);
 Bucket makeSpiBucket(document::BucketId bucketId, PartitionId partitionId);
 Bucket makeSpiBucket(document::BucketId bucketId);
