+++ searchcore/src/tests/proton/documentdb/document_subdbs/document_subdbs_test.cpp
@@ -1,6 +1,6 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
-#include <vespa/persistence/spi/test.h>
+#include <vespa/document/test/make_bucket_space.h>
 #include <vespa/searchcore/proton/attribute/imported_attributes_repo.h>
 #include <vespa/searchcore/proton/bucketdb/bucketdbhandler.h>
 #include <vespa/searchcore/proton/common/hw_info.h>
@@ -44,6 +44,7 @@ using namespace search;
 using namespace searchcorespi;
 using namespace vespalib;
 
+using document::test::makeBucketSpace;
 using proton::bucketdb::BucketDBHandler;
 using proton::bucketdb::IBucketDBHandler;
 using proton::bucketdb::IBucketDBHandlerInitializer;
@@ -52,7 +53,6 @@ using search::test::DirectoryHandler;
 using searchcorespi::IFlushTarget;
 using searchcorespi::index::IThreadingService;
 using storage::spi::Timestamp;
-using storage::spi::test::makeBucketSpace;
 using vespa::config::search::core::ProtonConfig;
 using vespalib::mkdir;
